Properly implement memory profiler
